add a cumulus image with management vrf enabled
Because enabling the management vrf breaks ssh connections and
consequently automation tools that use ssh (like Ansible) it is fairly
difficult to enable the management vrf in an automated way without
superflous hacking. So we are adding an image that starts life with the
management vrf enabled. The vrf is enabled through packers bootcmd
mechanism over vnc so it does not interfere with the ssh based packer
provisioning process.